requirement.txt
PyQt5
opencv-python
google-generativeai
qasync
Pillow
SpeechRecognition
pyttsx3